# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Video Topic Graph Platform API
  description: |
    Open-source platform for video topic graph generation.
    All endpoints require OIDC authentication via Bearer token.
  version: 1.0.0
  contact:
    name: Video Topic Graph Platform
    url: https://github.com/video-topic-graph/platform
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
  /videos/analyze:
    post:
      summary: Submit a video for analysis
      description: |
        Submits a video URL for processing through the complete pipeline:
        download → ASR → topic segmentation → graph generation → snippet creation
      tags: [Videos]
      operationId: analyzeVideo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoAnalyzeRequest'
      responses:
        '202':
          description: Video accepted for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
  /videos:
    get:
      summary: List user's videos
      tags: [Videos]
      operationId: listVideos
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/VideoStatus'
      responses:
        '200':
          description: List of videos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoListResponse'
  /videos/{id}:
    get:
      summary: Get video details
      tags: [Videos]
      operationId: getVideo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Video details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a video and all associated data
      tags: [Videos]
      operationId: deleteVideo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Video deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /videos/{id}/status:
    get:
      summary: Get video processing status
      tags: [Videos]
      operationId: getVideoStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Processing status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoStatusResponse'
  /videos/{id}/transcript:
    get:
      summary: Get video transcript segments
      tags: [Videos]
      operationId: getVideoTranscript
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: start_ts
          in: query
          schema:
            type: number
          description: Filter segments starting after this timestamp
        - name: end_ts
          in: query
          schema:
            type: number
          description: Filter segments ending before this timestamp
      responses:
        '200':
          description: Transcript segments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC Bearer token from Keycloak
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    VideoAnalyzeRequest:
      type: object
      required: [source_url]
      properties:
        source_url:
          type: string
          format: uri
          description: URL of the video to analyze
        source_type:
          type: string
          enum: [youtube, vimeo, direct, file]
          default: direct
        config:
          $ref: '#/components/schemas/PipelineConfig'
        metadata:
          type: object
          additionalProperties: true
          description: Optional user metadata
    PipelineConfig:
      type: object
      description: Override default pipeline configuration
      properties:
        asr_model:
          type: string
          enum: [whisper, faster-whisper, whisper-cpp]
        asr_language:
          type: string
          description: ISO 639-1 language code (auto-detect if omitted)
        enable_diarization:
          type: boolean
          default: false
        enable_scene_detection:
          type: boolean
          default: false
        topic_levels:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
        embedding_model:
          type: string
          enum: [all-MiniLM-L6-v2, all-mpnet-base-v2, ollama-nomic]
        llm_model:
          type: string
          enum: [llama2, mistral, mixtral, ollama-phi]
        snippet_quality:
          type: string
          enum: [low, medium, high]
          default: medium
...
    VideoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        source_url:
          type: string
          format: uri
        source_type:
          type: string
        status:
          $ref: '#/components/schemas/VideoStatus'
        duration_s:
          type: number
          nullable: true
        created_at:
          type: string
          format: date-time
        job_id:
          type: string
          format: uuid
          nullable: true
          description: ID of the processing job
    VideoDetailResponse:
      allOf:
        - $ref: '#/components/schemas/VideoResponse'
        - type: object
          properties:
            metadata:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                thumbnail_url:
                  type: string
                  format: uri
            graph:
              $ref: '#/components/schemas/GraphSummary'
            transcript_segments_count:
              type: integer
            topics_count:
              type: integer
            versions_count:
              type: integer
    VideoListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VideoResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    GraphSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        status:
          type: string
        nodes_count:
          type: integer
        edges_count:
          type: integer
...
    TranscriptResponse:
      type: object
      properties:
        video_id:
          type: string
          format: uuid
        segments:
          type: array
          items:
            $ref: '#/components/schemas/TranscriptSegment'
        total:
          type: integer

    # Entity Schemas
    VideoStatus:
      type: string
      enum:
        - pending
        - downloading
        - processing
        - analyzing
        - generating_graph
        - completed
        - failed
        - cancelled
    VideoStatusResponse:
      type: object
      properties:
        video_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/VideoStatus'
        progress:
          type: number
          minimum: 0
          maximum: 100
        current_stage:
          type: string
        stages:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              status:
                type: string
              progress:
                type: number
        error:
          type: string
          nullable: true
        estimated_completion:
          type: string
          format: date-time
          nullable: true
...
    TranscriptSegment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        start_ts:
          type: number
        end_ts:
          type: number
        speaker_id:
          type: string
          nullable: true
        text:
          type: string
        words:
          type: array
          items:
            $ref: '#/components/schemas/WordAlignment'
          nullable: true
    WordAlignment:
      type: object
      properties:
        word:
          type: string
        start_ts:
          type: number
        end_ts:
          type: number
        confidence:
          type: number
          nullable: true
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        request_id:
          type: string